DataUnions: encapsulate contract handling, remove obsolete wrappers, improve validation, memberStats format (#206)
Split DataUnionEndpoints.ts: move contract handling to Contracts.ts, ABI definitions to abi.ts and inline public methods to DataUnion.ts

Smaller changes:
- Remove obsolete caching from address fetching
- Improve address validation (and add tests)
- Split tests, improve test cleanup
- Bug fix for payForSignatureTransport option in withdraws: method option overrides client option
- Format of getMemberStats() data modified (enum values changed, number values string->BigNumber)
- Type annotations for StreamrClientOptions: many of the fields are required, as we apply a default value -> users give options as Partial<StreamrClientOptions>
- Id and debug are client properties, not options
