+++ package.json
@@ -65,7 +65,6 @@
     "@types/uuid": "^8.3.0",
     "@typescript-eslint/eslint-plugin": "^4.15.1",
     "@typescript-eslint/parser": "^4.15.1",
-    "async-mutex": "^0.3.0",
     "babel-loader": "^8.2.2",
     "babel-plugin-add-module-exports": "^1.0.4",
     "babel-plugin-transform-class-properties": "^6.24.1",
@@ -114,6 +113,7 @@
     "@ethersproject/sha2": "^5.0.8",
     "@ethersproject/transactions": "^5.0.10",
     "@ethersproject/wallet": "^5.0.11",
+    "@ethersproject/web": "^5.0.13",
     "debug": "^4.3.2",
     "eventemitter3": "^4.0.7",
     "lodash.uniqueid": "^4.0.1",
