+++ src/Config.ts
@@ -1,65 +1,55 @@
 import qs from 'qs'
-import Debug from 'debug'
 import { ControlLayer, MessageLayer } from 'streamr-client-protocol'
 import { ExternalProvider, JsonRpcFetchFunc } from '@ethersproject/providers'
 import { BigNumber } from '@ethersproject/bignumber'
-import { O } from 'ts-toolbelt'
-import { getVersionString, counterId } from './utils'
+import { getVersionString } from './utils'
+import { ConnectionInfo } from '@ethersproject/web'
 import { Todo } from './types'
 
 export type EthereumConfig = ExternalProvider|JsonRpcFetchFunc
 
 export type StreamrClientOptions = {
-    id?: string
-    debug?: Debug.Debugger,
-    auth?: {
+    auth: {
         privateKey?: string
         ethereum?: EthereumConfig
         apiKey?: string
         username?: string
         password?: string
     }
-    url?: string
-    restUrl?: string
-    streamrNodeAddress?: string
-    autoConnect?: boolean
-    autoDisconnect?: boolean
-    orderMessages?: boolean,
-    retryResendAfter?: number,
-    gapFillTimeout?: number,
-    maxGapRequests?: number,
-    maxPublishQueueSize?: number,
-    publishWithSignature?: Todo,
-    verifySignatures?: Todo,
-    publisherStoreKeyHistory?: boolean,
-    groupKeys?: Todo
-    keyExchange?: Todo
-    mainnet?: Todo
-    sidechain?: {
-        url?: string
-    },
+    url: string
+    restUrl: string
+    streamrNodeAddress: string
+    autoConnect: boolean
+    autoDisconnect: boolean
+    orderMessages: boolean
+    retryResendAfter: number
+    gapFillTimeout: number
+    maxGapRequests: number
+    maxPublishQueueSize: number
+    publishWithSignature: Todo
+    verifySignatures: Todo
+    publisherStoreKeyHistory: boolean
+    groupKeys: Todo
+    keyExchange: Todo
+    mainnet?: ConnectionInfo|string
+    sidechain?: ConnectionInfo|string
     dataUnion?: string
-    tokenAddress?: string,
-    minimumWithdrawTokenWei?: BigNumber|number|string,
-    sidechainTokenAddress?: string
-    factoryMainnetAddress?: string
-    factorySidechainAddress?: string
-    payForSignatureTransport?: boolean
-    cache?: {
-        maxSize?: number,
-        maxAge?: number
+    tokenAddress: string,
+    minimumWithdrawTokenWei?: BigNumber|number|string
+    factoryMainnetAddress: string
+    factorySidechainAddress: string
+    payForSignatureTransport: boolean
+    cache: {
+        maxSize: number,
+        maxAge: number
     }
 }
 
-export type StreamrClientConfig = O.Compulsory<StreamrClientOptions, 'url' | 'restUrl'>
 const { ControlMessage } = ControlLayer
 const { StreamMessage } = MessageLayer
 
 export default function ClientConfig(opts: Partial<StreamrClientOptions> = {}) {
-    const { id = counterId('StreamrClient') } = opts
-
-    const defaults = {
-        debug: Debug(id),
+    const defaults: StreamrClientOptions = {
         // Authentication: identity used by this StreamrClient instance
         auth: {}, // can contain member privateKey or (window.)ethereum
 
@@ -87,16 +77,11 @@ export default function ClientConfig(opts: Partial<StreamrClientOptions> = {}) {
         // Ethereum and Data Union related options
         // For ethers.js provider params, see https://docs.ethers.io/ethers.js/v5-beta/api-providers.html#provider
         mainnet: undefined, // Default to ethers.js default provider settings
-        sidechain: {
-            url: undefined, // TODO: add our default public service sidechain node, also find good PoA params below
-            // timeout:
-            // pollingInterval:
-        },
+        sidechain: undefined, // TODO: add our default public service sidechain node, also find good PoA params below
         tokenAddress: '0x0Cf0Ee63788A0849fE5297F3407f701E122cC023',
         minimumWithdrawTokenWei: '1000000', // Threshold value set in AMB configs, smallest token amount to pass over the bridge
-        sidechainTokenAddress: undefined, // TODO // sidechain token
-        factoryMainnetAddress: undefined, // TODO // Data Union factory that creates a new Data Union
-        factorySidechainAddress: undefined,
+        factoryMainnetAddress: 'TODO', // TODO // Data Union factory that creates a new Data Union
+        factorySidechainAddress: 'TODO',
         payForSignatureTransport: true, // someone must pay for transporting the withdraw tx to mainnet, either us or bridge operator
         cache: {
             maxSize: 10000,
@@ -104,7 +89,7 @@ export default function ClientConfig(opts: Partial<StreamrClientOptions> = {}) {
         }
     }
 
-    const options: StreamrClientConfig = {
+    const options: StreamrClientOptions = {
         ...defaults,
         ...opts,
         cache: {
