+++ test/integration/dataunion/calculate.test.ts
@@ -3,9 +3,9 @@ import debug from 'debug'
 
 import StreamrClient from '../../../src/StreamrClient'
 import config from '../config'
+import { createClient, expectInvalidAddress } from '../../utils'
 
-const log = debug('StreamrClient::DataUnionEndpoints::integration-test-calculate')
-// const { log } = console
+const log = debug('StreamrClient::DataUnion::integration-test-calculate')
 
 // @ts-expect-error
 const providerSidechain = new providers.JsonRpcProvider(config.clientOptions.sidechain)
@@ -16,28 +16,36 @@ const adminWalletMainnet = new Wallet(config.clientOptions.auth.privateKey, prov
 // This test will fail when new docker images are pushed with updated DU smart contracts
 // -> generate new codehashes for getDataUnionMainnetAddress() and getDataUnionSidechainAddress()
 
-it('DataUnionEndPoints: calculate DU address before deployment', async () => {
-    log(`Connecting to Ethereum networks, config = ${JSON.stringify(config)}`)
-    const network = await providerMainnet.getNetwork()
-    log('Connected to "mainnet" network: ', JSON.stringify(network))
-    const network2 = await providerSidechain.getNetwork()
-    log('Connected to sidechain network: ', JSON.stringify(network2))
+describe('DataUnion calculate', () => {
 
-    const adminClient = new StreamrClient(config.clientOptions as any)
-    await adminClient.ensureConnected()
+    afterAll(() => {
+        providerMainnet.removeAllListeners()
+        providerSidechain.removeAllListeners()
+    })
 
-    const dataUnionName = 'test-' + Date.now()
-    // eslint-disable-next-line no-underscore-dangle
-    const dataUnionPredicted = adminClient._getDataUnionFromName({ dataUnionName, deployerAddress: adminWalletMainnet.address })
+    it('calculate DU address before deployment', async () => {
+        log(`Connecting to Ethereum networks, config = ${JSON.stringify(config)}`)
+        const network = await providerMainnet.getNetwork()
+        log('Connected to "mainnet" network: ', JSON.stringify(network))
+        const network2 = await providerSidechain.getNetwork()
+        log('Connected to sidechain network: ', JSON.stringify(network2))
 
-    const dataUnionDeployed = await adminClient.deployDataUnion({ dataUnionName })
-    const version = await dataUnionDeployed.getVersion()
+        const adminClient = new StreamrClient(config.clientOptions as any)
 
-    await providerMainnet.removeAllListeners()
-    await providerSidechain.removeAllListeners()
-    await adminClient.ensureDisconnected()
+        const dataUnionName = 'test-' + Date.now()
+        // eslint-disable-next-line no-underscore-dangle
+        const dataUnionPredicted = adminClient._getDataUnionFromName({ dataUnionName, deployerAddress: adminWalletMainnet.address })
 
-    expect(dataUnionPredicted.getAddress()).toBe(dataUnionDeployed.getAddress())
-    expect(dataUnionPredicted.getSidechainAddress()).toBe(dataUnionDeployed.getSidechainAddress())
-    expect(version).toBe(2)
-}, 60000)
+        const dataUnionDeployed = await adminClient.deployDataUnion({ dataUnionName })
+        const version = await dataUnionDeployed.getVersion()
+
+        expect(dataUnionPredicted.getAddress()).toBe(dataUnionDeployed.getAddress())
+        expect(dataUnionPredicted.getSidechainAddress()).toBe(dataUnionDeployed.getSidechainAddress())
+        expect(version).toBe(2)
+    }, 60000)
+
+    it('get DataUnion: invalid address', () => {
+        const client = createClient(providerSidechain)
+        return expectInvalidAddress(async () => client.getDataUnion('invalid-address'))
+    })
+})
